ModulesDownload, KOfamDownload in kegg.py -> metabolism/downloads.py
